requirements.txt
urllib,
spacy,
pandas,
numpy,
sklearn.feature_extraction.text,
sklearn.metrics.pairwise,
sklearn.preprocessing,
sklearn.svm